Refactoring postgraphile_introspection and all tests
There is currently an error when I try to use the Postgres client in
order to inspect the database. Looking at the stack trace I found that
the Postgres client is just a wrapper on the async version of the
Client. I think that it tries to launch a secondary tokio async runtime